requirements.txt
pandas
numpy
pyyaml
typer
pandera
pytest
